requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]>=4.6
pymongo
scikit-learn

## This -e . will refer to the setup.py file in the current directory
## Write this at the last of the project while deploying
# -e .